config: don't skip image pull if url has changed
Let's say we have the following image plan defined:

```
my_image:
  type: image
  url: https://my-hosting-server.net/image_v1.qcow2
```

If later I decide to modify it to the following:

```
my_image:
  type: image
  url: https://my-hosting-server.net/image_v2.qcow2
```

`kcli` will skip pulling the new image if I delete my plan and recreate it
because it only relies on the fact the key 'my_image' is present in the
profile file (~/.kcli/profiles.yml).

The idea here is to check the filename of the image url defined in the
plan in addition to the current check.

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>