Remove --disable-network-policy
This is to make it possible for the end-user to decide whether or not the regular internal K3s network policy controller should be removed or not. As it's not necessarily the case one wants to remove that controller if it was decided to go away from **flannel**. E.g. not all CNI's may come with their own network policy controller.

Also as the `--disable-network-policy` was _already_ not on the bootstrap part of the create() func. this also aligns "things".