Brushups on K3s worker/agents. Network related K3s binary args (#323)
* Use proper arg. on workers when NOT using **flannel**

Just as we do on the bootstrapping part in the create() func. also use `--flannel-backend=none` for the worker setup in the create() func.

* Remove --disable-network-policy

This is to make it possible for the end-user to decide whether or not the regular internal K3s network policy controller should be removed or not. As it's not necessarily the case one wants to remove that controller if it was decided to go away from **flannel**. E.g. not all CNI's may come with their own network policy controller.

Also as the `--disable-network-policy` was _already_ not on the bootstrap part of the create() func. this also aligns "things".

* Worker nodes shouldn't get network related settings

As worker nodes/agents in K3s gets their network settings from masters network settings shouldn't be propagated to the workers/agents.

See [this](https://rancher-users.slack.com/archives/CGGQEHPPW/p1629227659194400?thread_ts=1629225486.193200&cid=CGGQEHPPW) comments from Mr. Brad Davidson from RancherLabs: `agents get their network config from the servers`

⚡ 👍