Update masters.sh
Do not remove`cURL` as it is used inside the keepalived check script - https://github.com/karmab/kcli/blob/0453417a8296f16a1dc9143841e9e64456bc1471/kvirt/k3s/keepalived.conf#L2 ... so for now we need to **NOT** uninstall `cURL` .... if we think it's a security liability to have ´cURL` on a node then we have to revisit this later on. E.g. maybe `nc` can be  used instead. I think that's always on a Linux distro per default.