Do not remove cURL for K3s installs (#396)
* Update workers.yml

Undo remove from `cURL` in the workers.yml file as well.

* Update bootstrap.sh

Do not remove `cURL`

* Update masters.sh

Do not remove cURL as it is used inside the keepalived check script -

kcli/kvirt/k3s/keepalived.conf

Line 2 in 0453417
 script "/usr/bin/timeout 0.9 /usr/bin/curl -o /dev/null -kLs https://0:6443/readyz" 
... 

so for now we need to NOT uninstall cURL .... if we think it's a security liability to have `cURL` on a node then we have to revisit this later on. E.g. maybe `nc` can be used instead. I think that's always on a Linux distro per default.