Removing the gcp-hack
It should really not be necessary to have this gcp-hack. Actually the hack, by removing the local route to the load balancer API IP, on the extra control-plane nodes, I was **NOT** able to communicate properly with the API IP in a local shell when exported to a GCP self-managed K8s cluster.

As far as I can tell the local route to the load balancer is needed for communication to properly come through to the K8s control-plane.