Update __init__.py to remove the gcp_hack
The GCP hack should really not be needed. Whether or not the load balancer fronting the API IP / control-plane is external or not. Each node of the control-plane participating "behind" the API IP should have a local route. That makes it possible for each node to properly route packets to the API IP on the load balancer fronting it ....

- Simplifies the deployment of a self-managed K8s cluster on GCP ( GCE )
- Simplifies the KCLI code

---

Again when this gcp_hack ran when deploying a Kubernetes cluster I was not able to properly communicate with the API IP before adding the local route to the load balancer ip on the extra control-plane nodes.