Fix plan create `TypeError` (#639)
* Fix TypeError exception on plan creation

When creating a plan there are many instances where `config.plan`
returns None, which ends in an ugly exception being raised:

```
Traceback (most recent call last):
  File "/usr/bin/kcli", line 33, in <module>
    sys.exit(load_entry_point('kcli==99.0', 'console_scripts', 'kcli')())
             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/lib/python3.12/site-packages/kvirt/cli.py", line 5446, in cli
    args.func(args)
  File "/usr/lib/python3.12/site-packages/kvirt/cli.py", line 2110, in create_plan
    if 'result' in result and result['result'] == 'success':
       ^^^^^^^^^^^^^^^^^^
TypeError: argument of type 'NoneType' is not iterable
```

This patch fixes it by treating a `None` result in a plan creation as
being an error without a message.

* Consider success when ansible in plan does nothing

Ansible is skipped When creating a plan where the instances are skipped
because they already exist, and a failure exit code is returned by kcli.

This patch modifies it so that this is considered a success, since a
plan creation was requested and it is already there.