Use correct SSH user when waiting for VM boot
A few lines above this change an SSH connection is being established to
test the gathered IP. This uses a previously determined user name.

The SSH connection to check whether the VM boot finished then used the
hard-coded root user. A VM can be configured with a different user where
SSH connections to the root user aren't possible. For such VMs the setup
process is stuck at this step, repeatedly failing to connect via SSH.

This is fixed by using the declared `user` variable which is set and
already used by the code above.

The `user` variable could be `None` in case the previous loop exits due
to the wait timeout being reached. In this case the `ip` would also be
`None` so this change doesn't introduce an additional failure case.