KVM: Fix hardcoded default pool dir (#714)
Current KVM provider code has the directory of the default pool
hardcoded in a couple of places.

This is problematic if you have your default pool in a different
location.

In this patch we replace hardcoded directory (`/var/lib/libvirt/images`)
with the known path of the default pool (`default_poolpath`).

Closes: #713