add kcli ssh -t pseudo-terminal allocation
When ssh(1) is invoked with a command it does not enable pty allocation.
The `ssh -t` option forces pty allocation and is typically used together
with commands that assume there is a terminal (for tab completion, job
control in shells, etc).

Add the `kcli ssh -t` option so that commands can be invoked with a
pseudo-terminal.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>