Fix plan of plans that use cwd
Using plan files from the current working directory in a plan of plans
doesn't work as expected.

Assuming we have 3 plans: myplan.yaml, libvirt.yaml, and ocp.yaml all in
the same directory.

Contents of `myplan.yaml` are:

```yaml
libvirt:
  type: plan
    file: libvirt.yaml

ocp:
  type: plan
    file: ocp.yaml
```

Then we'll get error

```
Input file libvirt not found.Leaving....
```

If we explicitly say files are in the current work directory with `./`
like this:

```
libvirt:
  type: plan
    file: ./libvirt.yaml

ocp:
  type: plan
    file: ./ocp.yaml
```

The the `libvirt.yaml` will be properly read, but the `ocp.yaml` won't
and we get error:

```
Input file ocp.yaml not found.Leaving....
```

This second error is caused by kcli starting the execution of the second
plan from the parent directory instead of where the execution was
started.

This patch fixes these 2 issues.