OpenShift: Web server on different IP
When serving ISO images for baremetal machines on the OpenShift cluster
deployment we were assuming that these were always served on the
`default` network interface, and that's the IP address that was
configured on the BMC of the baremetal server.

This only works if we are running kcli on the same network as the
servers, but that's not the case when we access them through a VPN.

In this patch we add a new configuration option named
`baremetal_web_nic` that allows us to specify the network interface we
want to use to serve the ISO.

The default value will be `default`, to maintain backward compatibility,
but for example we could set it to `tun0` when using the VPN to access
the machines to be provisioned.