Raise right exception on `set_iso`
There are a couple of places in the `set_iso` method where an exception
is not raised, so later on an invalid exception is raised, such as:

```
Hit 'NoneType' object has no attribute 'code' when plugging iso http://10.45.225.105/sno-devin-sno.iso to host with url https://devin03.mgmt.lab.eng.brq2.redhat.com/redfish/v1/Systems/System.Embedded.1
```

That `NoneType` is referring to the code trying `result.code` when
`result` has not been changed from the default because an exception had
been risen.

This patch fixes this error by raising the exception when it's
encountered.