Add destdir override in workflow (#786)
Existing workflows default to the `/root` as the parent destination directory
for a workflow's files.

This is problematic if we are running the workflow locally as a normal user,
for example with something like:

```
myworkflow:
  type: workflow
  scripts:
  - archs/create.sh
  files:
  - archs/poct/myapp.yaml
```

Because we'll get the following error:

```
PermissionError: [Errno 13] Permission denied: '/root/archs'
```

This doesn't have a nice solution based as we have the following constraints:

- Can only provide the `outputdir` parameter as a command line argument when
  creating a workflow, not when creating a plan that has the workflow.

- If we want to be backward compatible we cannot change the default `/root`
  directory to use the same `TemporaryDirectory()` that the `scripts` section
  uses by default.

- We can use a dictionary in `files` with the `path` key to indicate a
  destination directory, but that won't match the `TemporaryDirectory()`
  directory of the scripts.

The proposed solution is to add a `destdir` parameter in the workflow to define
the `outputdir`. It accepts relative directories and is overridden by the
`outputdir` command line argument:

So we could solve the issue with the following workflow:

```
myworkflow:
  type: workflow
  destdir: outdir
  scripts:
  - archs/create.sh
  files:
  - archs/poct/myapp.yaml
```