OpenShift: EFI clean old RHCOS entries
If we use the same host to redeploy an OpenShift cluster we'll end up
with a long list of RHCOS entries in EFI.

With this patch we clean not only the DVD and CD entry but also any
RHCOS entry previously created.

This way only the newly RHCOS created entry will be left.