OpenShift: EFI support more than 10 entries
Current code in `sno-finish.sh` assumes there can never be more than 10
entries, which is not correct.

If we've done a lot of retries deploying OpenShift on the same node we
may have more than 10 EFI entries.

With this patch we change the regex used to get the boot number as an
hexadecimal 4 digits entry.