OpenShift: Improve EFI code in sno-finish.sh (#789)
* OpenShift: EFI clean old RHCOS entries

If we use the same host to redeploy an OpenShift cluster we'll end up
with a long list of RHCOS entries in EFI.

With this patch we clean not only the DVD and CD entry but also any
RHCOS entry previously created.

This way only the newly RHCOS created entry will be left.

* OpenShift: EFI don't pass unnecessary $NUM

When cleaning EFI entries we are using:

```
efibootmgr -b 000$NUM -B $NUM
```

But according to the documentation the last $NUM is not necessary:

```
       -b | --bootnum XXXX
              Modify BootXXXX (hex).

       -B | --delete-bootnum
              Delete bootnum.
```

So we just remove it.

* OpenShift: EFI support more than 10 entries

Current code in `sno-finish.sh` assumes there can never be more than 10
entries, which is not correct.

If we've done a lot of retries deploying OpenShift on the same node we
may have more than 10 EFI entries.

With this patch we change the regex used to get the boot number as an
hexadecimal 4 digits entry.