App: Change targetNamespaces based on installModes
Currently whenever an operator defines a preferred installation
namespace we limit it's installation mode to that specific namespace of
the cluster.

This goes against the default behavior of the web console, where the
default (if the operator supports it) is to install it for all
namespaces of the cluster.

In this patch we change kcli to behave like the web console, so even if
the operator is installed in the desired namespace it will only be
restricted to that namespace if it doesn't support `AllNamespaces`.

This is important because there are operators, like the openstack
operator, that expects them to handle all the namespaces.

When that's not the case the installation of OpenStack in OpenShift
(following deployment guides) will not work, and we'll get the following
error when creating a valid `OpenStackControlPlane` object in the
`openstack` namespace:

```
The OpenStackControlPlane "rhoso" is invalid:
* spec.galera.templates.openstack.storageClass: Required value
* spec.galera.templates.openstack-cell1.storageClass: Required value
* spec.glance.template.glanceAPIs.default.containerImage: Required value
* spec.glance.template.imageCache: Required value
* spec.barbican.template.serviceAccount: Required value
* spec.nova.template.apiContainerImageURL: Required value
* spec.nova.template.computeContainerImageURL: Required value
* spec.nova.template.conductorContainerImageURL: Required value
* spec.nova.template.metadataContainerImageURL: Required value
* spec.nova.template.novncproxyContainerImageURL: Required value
* spec.nova.template.schedulerContainerImageURL: Required value
* <nil>: Invalid value: "null": some validation rules were not checked because the object was invalid; correct the existing errors to complete validation
```

This is because the mutating webhooks are restricted by the
`targetNamespaces`.