Fix workflow files without path
On commit f909009e669e8ca926ff2fa1713ace9ce8cf7b54 the workflow code was
cleaned up, but it broke files that had no `path`.

Error shown is similar to this:

```
    destdir = f"{destdir}/{os.path.dirname(entry['origin'])}"
                 ^^^^^^^
UnboundLocalError: cannot access local variable 'destdir' where it is not associated with a value
```

That's because in the previously mentioned commit variable
`default_destdir` that was removed was accidentally changed to `destdir`
instead of `outputdir`.

This patch fixes this problem using the right variable.