WIP: Minor doc improvements (#798)
* Docs: tar is necessary to install

When deploying some of the clusters, like openshift for example, `tar`
command is necessary to be present, otherwise it will download and then
fail to extract it.

In this patch we add `tar` to the installation process in the
documentation.

* Docs: Ensure default pool exists

In some platforms the default pool may not be present after installing
libvirt or qemu-kvm, so kcli will have problems to run some things.

This patch adds a some text to the README.md to explain how to check and
create if it doesn't exist.