OpenStack: Fix auth using password
Using password to authenticate in the OpenStack provider will result in
an error:

```
  File "/home/geguileo/.local/lib/python3.9/site-packages/keystoneauth1/session.py", line 1118, in request
    raise exceptions.from_response(resp, method, url)
keystoneauth1.exceptions.http.Unauthorized: The request you have made requires authentication. (HTTP 401) (Request-ID: req-4a2631c9-5606-403b-a4aa-6c43c6c1ddff)
```

This is because we are always defaulting to token authentication by
setting the `auth_token` variable instead of leaving it empty, as we are
already marking it as password auth in the `default_auth_type`.

This patch fixes this.