Openstack certs (#800)
* OpenStack: Insecure connections

This patch adds the possibility of passing False or false in the
`ca_file` for the OpenStack provider to do insecure connections.

* OpenStack: Fix swift client

The Swift client rn the OpenStack provider doesn't use the session that
it receives during initialization, so it will ignore anything related to
the certificate (insecure and ca_file).

This patch fixes this passing the parameters explicitly.

Related upstream bug: https://bugs.launchpad.net/python-swiftclient/+bug/1736135

* OpenStack: self-signed cert on ignition

When using `ca_file` parameter for self-signed certificates with the
OpenStack provider the bootstraping node will fail with:

```
[  602.320900] ignition[1054]: GET https://swift-public-myurl.com/v1/AUTH_245a83566a284b0f8d618f3bbdbe94a0/sno1-lab-myurl-com/bootstrap.ign: attempt #122
[  602.368010] ignition[1054]: GET error: Get "https://swift-public-myurl.com/v1/AUTH_245a83566a284b0f8d618f3bbdbe94a0/sno1-lab-myurl-com/bootstrap.ign": tls: failed to verify certificate: x509: certificate signed by unknown authority
```

With this patch we add the custom certificate to the ignition file, so
it can connect to the remote location and get the ignition file for
merging.

We explicitly ignore ca_file set to `false` or `False` because I don't
know how to make ignition do insecure get.